% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterKmeans.R
\name{clusterKmeans.chan}
\alias{clusterKmeans.chan}
\title{K-means clustering}
\usage{
clusterKmeans.chan(x, k = 10, init.method = "pca-part", num.threads = 1)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells.}

\item{k}{Integer scalar specifying the number of clusters.}

\item{init.method}{String specifying the initialization method:
PCA partitioning (\code{"pca-part"}), kmeans++ (\code{"kmeans++"}) or random initialization (\code{"random"}).}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A list containing:
\itemize{
\item \code{clusters}, an integer vector containing the cluster assignments.
\item \code{centers}, a numeric matrix with the coordinates of the cluster centroids (dimensions in rows, centers in columns).
\item \code{iterations}, an integer scalar specifying the number of Hartigan-Wong iterations used.
\item \code{withinss}, a numeric vector containing the within-cluster sum of squares for each cluster.
}
}
\description{
Perform k-means clustering using kmeans++ initialization with the Hartigan-Wong algorithm.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
clustering <- clusterKmeans.chan(x)
table(clustering$clusters)

}
\author{
Aaron Lun
}
