% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSNNGraph.R
\name{clusterSNNGraph.chan}
\alias{clusterSNNGraph.chan}
\title{Apply graph-based clustering}
\usage{
clusterSNNGraph.chan(x, num.neighbors = 10, resolution = 1, num.threads = 1)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells.}

\item{num.neighbors}{Integer scalar specifying the number of neighbors to use to construct the graph.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A list containing:
\itemize{
\item \code{membership}, a list of integer vectors with cluster assignments for each cell at each level.
Assignments are sorted by decreasing resolution (i.e., fewer, larger clusters).
\item \code{modularity}, a numeric vector containing the modularity of each level.
\item \code{best}, the level with the lowest modularity.
}
}
\description{
Perform multi-level (i.e., Louvain) clustering on a shared nearest neighbor graph.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
clustering <- clusterSNNGraph.chan(x)
clustering$modularity
lapply(clustering$membership, table)

}
\author{
Aaron Lun
}
