% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMatrix.R
\name{combineMatrix}
\alias{combineMatrix}
\alias{rbindMatrix}
\alias{cbindMatrix}
\title{Combine multiple initialized matrices}
\usage{
rbindMatrix(...)

cbindMatrix(...)
}
\arguments{
\item{...}{Any number of lists of matrix data like that produced by \code{\link{initializeSparseMatrix}}.
All of them should have the same number of columns (for \code{rbindMatrix}) or columns (for \code{cbindMatrix}).}
}
\value{
A list containing the combined matrix data, along with their combined dimnames.
}
\description{
Combine multiple initialized matrices
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(100, 10, 0.1) * 100))
rownames(x) <- sprintf("GENE_\%i", seq_len(100))
colnames(x) <- LETTERS[1:10]

y <- initializeSparseMatrix(x)
ry <- rbindMatrix(y, y)
cy <- cbindMatrix(y, y)

}
\author{
Aaron Lun
}
