% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedSizeFactors.R
\name{groupedSizeFactors}
\alias{groupedSizeFactors}
\title{Compute grouped size factors}
\usage{
groupedSizeFactors(
  x,
  group,
  center = TRUE,
  prior.count = 10,
  reference = NULL,
  size.factors = NULL,
  batch = NULL,
  num.comp = 25,
  num.clusters = 20,
  num.threads = 1
)
}
\arguments{
\item{x}{A list of matrix data like that produced by \code{\link{initializeSparseMatrix}}.}

\item{group}{Vector or factor of length equal to the number of columns of \code{x},
containing the group assignment for each cell in \code{x}.
Alternatively \code{NULL}, in which case k-means clustering is performed internally.}

\item{center}{Logical scalar specifying whether the size factors should be centered on output.}

\item{prior.count}{Numeric scalar specifying the prior count to add to the pseudo-bulk profiles.
Larger values improve stability at the cost of accurate removal of composition biases.}

\item{reference}{Identity of the group in \code{group} to be used as a reference.
If \code{NULL}, a suitable reference is automatically chosen.}

\item{size.factors}{Passed to \code{\link{logNormCounts.chan}}.
Only used if \code{group=NULL}.}

\item{batch}{Passed to \code{\link{logNormCounts.chan}} and \code{\link{runPCA.chan}}.
Only used if \code{group=NULL}.}

\item{num.comp}{Number of principal components to use for clustering.
Only relevant if \code{group=NULL}.}

\item{num.clusters}{Number of k-means clusters to generate.
Only used if \code{group=NULL}.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
Numeric vector containing the size factor for each cell.
}
\description{
Correct for composition biases across groups of cells by computing size factors from per-group pseudo-bulk profiles.
}
\details{
If \code{group=NULL}, we generate a k-means clustering on the top PCs of the log-normalized matrix generated from \code{x}.
This involves:
\itemize{
\item Running \code{\link{logNormCounts.chan}} on \code{x}.
\item Running \code{\link{runPCA.chan}} to obtain \code{num.comp} PCs.
\item Running \code{\link{clusterKmeans.chan}} to obtain \code{num.clusters} clusters.
}
If \code{batch} is provided, this is used in \code{\link{logNormCounts.chan}} and \code{\link{runPCA.chan}} to reduce the impact of large/deeply sequenced batches.
}
\examples{
library(Matrix)
x <- round(abs(rsparsematrix(1000, 100, 0.1) * 100))
y <- initializeSparseMatrix(x)
groupedSizeFactors(y, group=NULL)

}
\author{
Aaron Lun
}
