% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSparseMatrix.R
\name{initializeSparseMatrix}
\alias{initializeSparseMatrix}
\title{Initialize the count matrix}
\usage{
initializeSparseMatrix(x, num.threads = 1)
}
\arguments{
\item{x}{A matrix-like object.
This should contain non-negative integers.}
}
\value{
A list of internal components that can be used in other \pkg{scran.chan} functions.
}
\description{
Initialize a matrix object on the C++ side.
If the input is a \linkS4class{dgCMatrix}, initialization is performed without copying any data;
otherwise an efficient block realization is performed.
}
\details{
Do not attempt to serialize this object; it contains a pointer to external memory, and will not be valid after a save/load cycle.
}
\examples{
# Mocking up a count matrix:
x <- Matrix::rsparsematrix(1000, 100, 0.1)
y <- round(abs(x))

stuff <- initializeSparseMatrix(y)
str(stuff)

# Trying with a DelayedArray:
z <- DelayedArray::DelayedArray(y)
stuff2 <- initializeSparseMatrix(z)
str(stuff2)

}
