% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNormCounts.R
\name{logNormCounts.chan}
\alias{logNormCounts.chan}
\title{Log-normalize the counts}
\usage{
logNormCounts.chan(x, size.factors = NULL, batch = NULL)
}
\arguments{
\item{x}{A list of matrix data like that produced by \code{\link{initializeSparseMatrix}}.}

\item{size.factors}{A numeric vector of length equal to the number of cells in \code{x},
containing positive size factors for all cells.}

\item{batch}{Vector or factor of length equal to the number of cells, specifying the batch of origin for each cell.
Alternatively \code{NULL} if all cells belong to the same batch.}
}
\value{
A list like \code{x} where the count matrix is replaced with a log-expression matrix.
}
\description{
Compute log-transformed normalized expression values by applying scaling normalization to count data.
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(1000, 100, 0.1) * 100))
y <- initializeSparseMatrix(x)
normed <- logNormCounts.chan(y)

}
\author{
Aaron Lun
}
