% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshalToSCE.R
\name{marshalToSCE}
\alias{marshalToSCE}
\title{Marshal results into a SingleCellExperiment}
\usage{
marshalToSCE(x, results, assay.type = "counts", include.gene.data = FALSE)
}
\arguments{
\item{x}{A SummarizedExperiment or SingleCellExperiment object containing the count matrix used in \code{\link{quickBasicAnalysis}}.}

\item{results}{A named list containing the output of \code{\link{quickBasicAnalysis}}.}

\item{assay.type}{String or integer scalar specifying the assay of \code{x} containing the counts.}

\item{include.gene.data}{Logical scalar indicating whether to include per-gene statistics in the \code{rowData}.}
}
\value{
A SingleCellExperiment decorated with size factors, reduced dimensions and cluster annotations.
}
\description{
Marshal the results of \code{\link{quickBasicAnalysis}} into a SingleCellExperiment object.
}
\details{
Size factors are stored in the \code{sizeFactors} of the output.
We also compute log-transformed normalized expression values and store them in the \code{"logcounts"} assay.

All dimensionality reduction results (PCA, t-SNE and UMAP) are stored in the \code{reducedDims}.

Clustering results are stored in \code{clusters} columns of the \code{colData}.
For multi-level community detection, a clustering result at level \code{N} is stored as \code{clusters.N}.
The clustering with the highest modularity is stored in the \code{colLabels}.

If \code{include.gene.data=TRUE}, per-gene statistics are stored as nested DataFrames inside the \code{rowData}.
Variance modelling statistics are stored in the \code{variances} column.
}
\examples{
library(scRNAseq)
x <- ZeiselBrainData()
mito <- grep("^mt-", rownames(x))
out <- quickBasicAnalysis(assay(x), qc.subsets=list(Mt=mito), num.threads=1)
sce <- marshalToSCE(x, out, assay.type=1)
sce

}
\author{
Aaron Lun
}
