% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnnCorrect.R
\name{mnnCorrect.chan}
\alias{mnnCorrect.chan}
\title{MNN correction}
\usage{
mnnCorrect.chan(
  x,
  batch,
  k = 15,
  nmads = 3,
  order = NULL,
  reference.policy = c("max-size", "max-variance", "max-rss", "input"),
  approximate = FALSE,
  num.threads = 1
)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells,
typically generated from \code{\link{runPCA.chan}}.}

\item{batch}{Vector or factor of length equal to the number of cells, specifying the batch of origin for each cell.}

\item{k}{Integer scalar specifying the number of neighbors to use when identifying MNN pairs.}

\item{nmads}{Numeric scalar specifying the number of MADs to use when computing correction vectors.}

\item{order}{Vector containing levels of \code{batch} in the desired merge order.
If \code{NULL}, a suitable merge order is automatically determined.}

\item{reference.policy}{String specifying the policy to use to choose the first reference batch.
This can be based on the largest batch (\code{"max-size"}), the most variable batch (\code{"max-variance"}), 
some combination of those two (\code{"max-rss"}) or the first specified input (\code{"input"}).
Only used for automatic merges, i.e., when \code{order=NULL}.}

\item{approximate}{Logical scalar specifying whether to perform an approximate neighbor search.
Set to \code{FALSE} for historical reasons.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
List containing:
\itemize{
\item \code{corrected}, a numeric matrix of the same dimensions as \code{x}, containing the corrected values.
\item \code{merge.order}, character vector containing the unique levels of \code{batch} in the automatically determined merge order.
The first level in this vector is used as the reference batch; all other batches are iteratively merged to it.
\item \code{num.pairs}, integer vector of length equal to the number of batches minus 1.
This contains the number of MNN pairs at each merge.
}
}
\description{
Apply mutual nearest neighbor (MNN) correction to remove batch effects from a low-dimensional matrix.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
b <- sample(1:2, ncol(x), replace=TRUE)
corrected <- mnnCorrect.chan(x, b)
str(corrected)

}
\author{
Aaron Lun
}
