% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perCellAdtQcMetrics.R
\name{perCellAdtQcMetrics.chan}
\alias{perCellAdtQcMetrics.chan}
\title{Compute per-cell ADT QC metrics}
\usage{
perCellAdtQcMetrics.chan(x, subsets, num.threads = 1)
}
\arguments{
\item{x}{List containing the output of \code{\link{initializeSparseMatrix}},
corresponding to a matrix of ADT counts.}

\item{subsets}{List of logical vectors specifying tag subsets of interest, typically control tags like IgGs.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A list containing:
\itemize{
\item \code{sum}, a numeric vector containing the total ADT count for each cell.
\item \code{detected}, an integer vector containing the number of detected tags per cell.
\item \code{subsets}, a list of numeric vectors containing the total count of each control set. 
}
Each vector is of length equal to the number of cells.
}
\description{
Calculate per-cell QC metrics from an initialized matrix of ADT counts.
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(1000, 100, 0.1) * 100))

# Mocking up a control set.
sub <- list(IgG=rbinom(nrow(x), 1, 0.1) > 0)

# Running the analysis:
y <- initializeSparseMatrix(x)
qc <- perCellAdtQcMetrics.chan(y, sub)
summary(qc$sum)
summary(qc$detected)
summary(qc$subsets$IgG)

}
\author{
Aaron Lun
}
