% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perCellCrisprQcMetrics.R
\name{perCellCrisprQcMetrics.chan}
\alias{perCellCrisprQcMetrics.chan}
\title{Compute per-cell CRISPR QC metrics}
\usage{
perCellCrisprQcMetrics.chan(x, num.threads = 1)
}
\arguments{
\item{x}{List containing the output of \code{\link{initializeSparseMatrix}},
corresponding to a matrix of CRISPR guide counts.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A list containing:
\itemize{
\item \code{sum}, a numeric vector containing the total tag count for each cell.
\item \code{detected}, an integer vector containing the number of detected tags per cell.
\item \code{max_proportion}, a numeric vector specifying the proportion of counts assigned to the most abundant tag in each cell.
\item \code{max_index}, an integer vector containing the index of the most abundant tag in each cell.
}
Each vector is of length equal to the number of cells.
}
\description{
Calculate per-cell QC metrics from an initialized matrix of CRISPR guide counts.
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(20, 100, 0.1) * 100))

y <- initializeSparseMatrix(x)
qc <- perCellCrisprQcMetrics.chan(y)
summary(qc$max_proportion)
table(qc$max_index)

}
\author{
Aaron Lun
}
