% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perCellQCMetrics.lite.R
\name{perCellQCMetrics.lite}
\alias{perCellQCMetrics.lite}
\title{Compute per-cell QC metrics}
\usage{
perCellQCMetrics.lite(x, subsets = NULL)
}
\arguments{
\item{x}{One of the supported matrix-like objects, containing genes as rows and cells as columns.}

\item{subsets}{List of logical vectors specifying feature subsets of interest.}
}
\value{
A list containing the total sum for each cell,
the number of detected features per cell,
and the proportion of counts in each feature subset.
}
\description{
Light-weight calculation of the per-cell QC metrics.
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(1000, 100, 0.1) * 100))

# Mocking up a subset:
sub <- list(Mito=rbinom(nrow(x), 1, 0.1) > 0)

qc <- diet.scran:::perCellQCMetrics.lite(x, sub)
summary(qc$sum)
summary(qc$detected)
summary(qc$subsets$Mito)

}
\author{
Aaron Lun
}
