% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectNeighborEmbedding.R
\name{projectNeighborEmbedding.chan}
\alias{projectNeighborEmbedding.chan}
\title{Project cells onto an embedding using its neighbors}
\usage{
projectNeighborEmbedding.chan(
  ref,
  embedding,
  test,
  k = 10,
  approximate = TRUE,
  num.threads = 1
)
}
\arguments{
\item{ref}{Numeric matrix for the reference dataset, where rows are dimensions and columns are cells.
This usually contains PC coordinates from \code{\link{runPCA.chan}} or friends.}

\item{embedding}{Numeric matrix for the embedding of the reference dataset, where rows are cells and columns are dimensions.}

\item{test}{Numeric matrix for the test dataset, where rows are dimensions and columns are cells.
This usually contains PC coordinates similar to \code{ref}.}

\item{k}{Integer scalar specifying the number of neighbors to use for projection.}

\item{approximate}{Logical scalar indicating whether an approximate neighbor search should be performed.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
Numeric matrix containing the projected embedding coordinates for each cell in the test dataset.
Like \code{embedding}, cells are in the rows.
}
\description{
Project cells into a reference embedding (e.g., t-SNE, UMAP) given its neighbors in another space.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
keep <- downsampleByNeighbors.chan(x, 3) # not too much downsampling
sub <- x[,keep]

embedding <- runUMAP.chan(sub)
full <- projectNeighborEmbedding.chan(sub, embedding, x)
plot(full[,1], full[,2], col=iris[,"Species"], 
    pch=ifelse(seq_len(ncol(x)) \%in\% keep, 16, 1)) 

}
\seealso{
\code{\link{downsampleByNeighbors.chan}}, which is typically used to generate a reference subsample for easier clustering before assignment of the full (test) dataset.
}
\author{
Aaron Lun
}
