% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickBasicAnalysis.R
\name{quickBasicAnalysis}
\alias{quickBasicAnalysis}
\title{Perform a simple scRNA-seq analysis}
\usage{
quickBasicAnalysis(
  x,
  qc.subsets,
  qc.nmads = 3,
  trend.span = 0.3,
  hvg.num = 2000,
  pca.num = 50,
  tsne.perplexity = 30,
  umap.num.neighbors = 15,
  cluster.snn.num.neighbors = 10,
  cluster.snn.resolution = 1,
  num.threads = 1
)
}
\arguments{
\item{x}{A sparse matrix-like object.}

\item{qc.subsets}{Feature control subsets, see \code{?\link{perCellQCMetrics.chan}} for details.}

\item{qc.nmads}{Filter threshold stringency, see \code{?\link{perCellQCFilters.chan}} for details.}

\item{trend.span}{Trend smoothing, see \code{?\link{modelGeneVar.chan}} for details.}

\item{hvg.num}{Integer scalar specifying the number of HVGs to use for PCA.}

\item{pca.num}{Number of top PCs, see \code{?\link{runPCA.chan}} for details.}

\item{tsne.perplexity}{Parameters to be used for t-SNE, see \code{\link{runTSNE.chan}} for details.}

\item{umap.num.neighbors}{Parameters to be used for UMAP, see \code{\link{runUMAP.chan}} for details.}

\item{cluster.snn.num.neighbors}{Parameters to be used for graph-based clustering, see \code{\link{clusterSNNGraph.chan}} for details.}

\item{cluster.snn.resolution}{Parameters to be used for graph-based clustering, see \code{\link{clusterSNNGraph.chan}} for details.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A list containing the elements:
\itemize{
    \item \code{qc.metrics}, a list containing the output of \code{\link{perCellQCMetrics.chan}}.
    \item \code{qc.filters}, a list containing the output of \code{\link{perCellQCFilters.chan}}.
    The \code{overall} vector is used to determine which cells to discard during quality control.
    \item \code{size.factors}, a numeric vector containing the size factors for scaling normalization.
    \item \code{variances}, a list containing the output of \code{\link{modelGeneVar.chan}}.
    This contains a \code{keep} column indicating whether the gene is retained for PCA.
    \item \code{pca}, a list containing the output of \code{\link{runPCA.chan}}.
    \item \code{cluster.snn}, a list containing the output of \code{\link{clusterSNNGraph.chan}}.
    \item \code{umap}, a list containing the output of \code{\link{runUMAP.chan}}.
    \item \code{tsne}, a list containing the output of \code{\link{runTSNE.chan}}.
}
}
\description{
Analyze a simple single-cell RNA-seq dataset, from quality control to marker gene detection.
No frills, no dependencies.
}
\examples{
library(scRNAseq)
x <- ZeiselBrainData()
mito <- grep("^mt-", rownames(x)))
out <- quickBasicAnalysis(assay(x), qc.subsets=list(Mt=x), num.threads=4)

}
\author{
Aaron Lun
}
