% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickMergedAnalysis.R
\name{quickMergedAnalysis}
\alias{quickMergedAnalysis}
\title{Perform a merged scRNA-seq analysis}
\usage{
quickMergedAnalysis(
  x,
  batch,
  qc.subsets,
  qc.nmads = 3,
  trend.span = 0.3,
  hvg.num = 2000,
  pca.num = 25,
  ...,
  correct.method = c("mnn", "regress"),
  markers.pref = c("snn", "kmeans"),
  num.threads = 1
)
}
\arguments{
\item{x}{A sparse matrix-like object.}

\item{qc.subsets}{Feature control subsets, see \code{?\link{perCellQCMetrics.chan}} for details.}

\item{qc.nmads}{Filter threshold stringency, see \code{?\link{perCellQCFilters.chan}} for details.}

\item{trend.span}{Trend smoothing, see \code{?\link{modelGeneVar.chan}} for details.}

\item{hvg.num}{Integer scalar specifying the number of HVGs to use for PCA.}

\item{pca.num}{Number of top PCs, see \code{?\link{runPCA.chan}} for details.}

\item{...}{Further arguments to pass to \code{\link{runAllDownstream}}.}

\item{markers.pref}{Character vector specifying the preference of clustering to use for marker calculation.
By default, if both graph-based and k-means clusters are available, markers are computed for the former;
otherwise it will use whatever clustering is available from the parameters in \code{...}.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A list containing the elements:
\itemize{
    \item \code{qc.metrics}, a list containing the output of \code{\link{perCellQCMetrics.chan}}.
    \item \code{qc.filters}, a list containing the output of \code{\link{perCellQCFilters.chan}}.
    \item \code{qc.discard}, a logical vector specifying which cells to discard during quality control.
    This is the same as \code{qc.filters$filters$overall}.
    \item \code{size.factors}, a numeric vector containing the size factors for scaling normalization.
    \item \code{variances}, a list containing the output of \code{\link{modelGeneVar.chan}}.
    This contains a \code{keep} column indicating whether the gene is retained for PCA.
    \item \code{pca}, a list containing the output of \code{\link{runPCA.chan}}.
}
In addition, some or all of the outputs of \code{\link{runAllDownstream}} may be present depending on the arguments in \code{...}.
By default, this includes:
\itemize{
    \item \code{cluster.snn}, a list containing the output of \code{\link{clusterSNNGraph.chan}}.
    \item \code{umap}, a list containing the output of \code{\link{runUMAP.chan}}.
    \item \code{tsne}, a list containing the output of \code{\link{runTSNE.chan}}.
}
}
\description{
Analyze a multi-batch single-cell RNA-seq dataset, from quality control to batch correction to marker gene detection.
}
\examples{
library(scRNAseq)
x <- ZeiselBrainData()
mito <- grep("^mt-", rownames(x))
out <- quickMergedAnalysis(assay(x), qc.subsets=list(Mt=mito), num.threads=1)

}
\author{
Aaron Lun
}
