% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAllDownstream.R
\name{runAllDownstream}
\alias{runAllDownstream}
\title{Run all downstream steps}
\usage{
runAllDownstream(
  x,
  do.tsne = TRUE,
  do.umap = TRUE,
  do.cluster.snn = TRUE,
  do.cluster.kmeans = FALSE,
  tsne.args = list(),
  umap.args = list(),
  cluster.snn.args = list(),
  cluster.kmeans.args = list(),
  tsne.perplexity = NULL,
  umap.num.neighbors = NULL,
  umap.min.dist = NULL,
  cluster.snn.num.neighbors = NULL,
  cluster.snn.method = NULL,
  cluster.snn.resolution = NULL,
  cluster.kmeans.k = NULL,
  cluster.kmeans.init = NULL,
  drop = TRUE,
  approximate = TRUE,
  num.threads = 1
)
}
\arguments{
\item{x}{Numeric matrix containing dimensions in the rows and cells in the columns.
This is typically a matrix of principal components.}

\item{do.tsne}{Logical scalar, should we perform a t-SNE?}

\item{do.umap}{Logical scalar, should we perform a UMAP?}

\item{do.cluster.snn}{Logical scalar, should we perform graph-based clustering?}

\item{do.cluster.kmeans}{Logical scalar, should we perform k-means clustering?}

\item{tsne.args}{Named list of t-SNE parameters, see \code{\link{runTSNE.chan}} for details.}

\item{umap.args}{Named list of UMAP parameters, see \code{\link{runUMAP.chan}} for details.}

\item{cluster.snn.args}{Named list of SNN parameters, see \code{\link{clusterSNNGraph.chan}} for details.}

\item{cluster.kmeans.args}{Named list of k-means parameters, see \code{\link{clusterKmeans.chan}} for details.}

\item{tsne.perplexity}{Deprecated, use \code{tsne.args} instead.}

\item{umap.num.neighbors, umap.min.dist}{Deprecated, use \code{umap.args} instead.}

\item{cluster.snn.num.neighbors, cluster.snn.method, cluster.snn.resolution}{Deprecated, use \code{cluster.snn.args} instead.}

\item{cluster.kmeans.k, cluster.kmeans.init}{Deprecated, use \code{cluster.kmeans.args} instead.}

\item{drop}{Logical scalar indicating whether to drop the sweep-based formatting when the parameters are scalars.}

\item{approximate}{Logical scalar specifying whether to perform an approximate neighbor search.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A list containing \code{"cluster.snn"}, \code{"cluster.kmeans"}, \code{"umap"} and \code{"tsne"},
each of which contains the result of their respective function \code{*.chan} functions.
}
\description{
Run all steps in the basic workflow that are downstream of PCA (but before marker detection).
}
\details{
By running all of these steps together, we can avoid redundant construction of the nearest neighbor index.
We can also execute some of the single-threaded steps concurrently for further time savings.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
res <- runAllDownstream(x, num.threads=3)
str(res)
}
\author{
Aaron Lun
}
