% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSNE.R
\name{runTSNE.chan}
\alias{runTSNE.chan}
\title{Run an approximate t-SNE}
\usage{
runTSNE.chan(
  x,
  perplexity = 30,
  interpolate = -1,
  max.depth = 7,
  num.threads = 1
)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells.}

\item{perplexity}{Numeric scalar specifying the perplexity to use in the t-SNE algorithm.}

\item{interpolate}{Integer scalar specifying the grid resolution for interpolating repulsive forces (larger is slower but more accurate).
A value of zero disables interpolation, while a value of -1 only uses interpolation for large datasets.}

\item{max.depth}{Integer scalar specifying the maximum depth of the Barnes-Hut quad trees (larger is slower but more accurate).
Set to a large integer (e.g., 1000) to eliminate depth restrictions.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A numeric matrix where rows are cells and columns are the two dimensions of the embedding.
}
\description{
Compute the t-stochastic neighbor embedding with assorted approximations for speed.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
embedding <- runTSNE.chan(x)
plot(embedding[,1], embedding[,2], col=iris[,5])

}
\author{
Aaron Lun
}
