% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUMAP.R
\name{runUMAP.chan}
\alias{runUMAP.chan}
\title{Run UMAP}
\usage{
runUMAP.chan(x, num.neighbors = 15, num.threads = 1)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells.}

\item{num.neighbors}{Integer scalar specifying the number of neighbors to use in the UMAP algorithm.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A numeric matrix where rows are cells and columns are the two dimensions of the embedding.
}
\description{
Compute the uniform manifold approximation and projection.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
embedding <- runUMAP.chan(x)
plot(embedding[,1], embedding[,2], col=iris[,5])

}
\author{
Aaron Lun
}
