% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleByNeighbors.R
\name{scaleByNeighbors.chan}
\alias{scaleByNeighbors.chan}
\title{Scale and combine multiple embeddings}
\usage{
scaleByNeighbors.chan(x, k = 20, approximate = TRUE, num.threads = 1)
}
\arguments{
\item{x}{List of numeric matrices of PCs or other embeddings where rows are dimensions and columns are cells.
All entries should have the same number of columns.}

\item{k}{Integer scalar specifying the number of neighbors to use.}

\item{approximate}{Logical scalar indicating whether an approximate neighbor search should be performed.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
List containing \code{scaling}, a vector of scaling factors to be aplied to each embedding;
and \code{combined}, a numeric matrix formed by scaling each entry of \code{x} and then \code{rbind}ing them together.
}
\description{
Scale multiple embeddings (usually derived from different modalities across the same set of cells) so that their within-population variances are comparable,
and then combine them into a single embedding matrix for combined downstream analysis.
}
\examples{
pcs <- list(
    gene = matrix(rnorm(10000), ncol=200),
    protein = matrix(rnorm(1000, sd=3), ncol=200),
    guide = matrix(rnorm(2000, sd=5), ncol=200)
)

out <- scaleByNeighbors.chan(pcs)
out$scaling
dim(out$combined)
}
\author{
Aaron Lun
}
