% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMarkers.R
\name{scoreMarkers.chan}
\alias{scoreMarkers.chan}
\title{Score marker genes}
\usage{
scoreMarkers.chan(x, groups)
}
\arguments{
\item{x}{A list of matrix data like that produced by \code{\link{logNormCounts}}.}

\item{groups}{A vector specifying the group assignment for each cell in \code{x}.}
}
\description{
Score marker genes using a variety of effect sizes from pairwise comparisons.
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(1000, 100, 0.1) * 100))
y <- initializeSparseMatrix(x)
normed <- logNormCounts.chan(y)

# Making up some groups.
g <- sample(4, ncol(x), replace=TRUE)
markers <- scoreMarkers.chan(normed, g)
names(markers)
head(markers[[1]])

}
