% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMatrix.R
\name{subsetMatrix}
\alias{subsetMatrix}
\title{Subset an initialized matrix}
\usage{
subsetMatrix(x, i, j)
}
\arguments{
\item{x}{A list of matrix data like that produced by \code{\link{initializeSparseMatrix}}.}

\item{i, j}{Subsetting vectors along the rows (\code{i}) or columns (\code{j}).
Each vector may be numeric, in which case it should contain indices;
logical, specifying a filter of length equal to the number of rows/columns;
or character, containing the names of the rows/columns to retain.
This can also be set to \code{NULL} to not perform filtering on that dimension.}
}
\value{
A list containing the subsetted matrix data, along with correspondingly subsetted dimnames.
}
\description{
Subset an initialized matrix
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(100, 10, 0.1) * 100))
rownames(x) <- sprintf("GENE_\%i", seq_len(100))
colnames(x) <- LETTERS[1:10]
y <- initializeSparseMatrix(x)

subsetMatrix(y, 1:5, c("A", "C", "E"))
subsetMatrix(y, 10:5, NULL)
}
\author{
Aaron Lun
}
