% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestRnaQcFilters.R
\name{suggestRnaQcFilters.chan}
\alias{suggestRnaQcFilters.chan}
\alias{perCellQCFilters.chan}
\title{Suggest filters for RNA QC metrics}
\usage{
suggestRnaQcFilters.chan(sums, detected, subsets, batch = NULL, nmads = 3)

perCellQCFilters.chan(sums, detected, subsets, batch = NULL, ...)
}
\arguments{
\item{sums}{Numeric vector containing the sum of counts for each cell.}

\item{detected}{Integer vector containing the total number of detected features for each cell.}

\item{subsets}{List of numeric vectors containing the proportion of counts assigned to each feature subset in each cell.}

\item{batch}{Vector or factor of length equal to the number of cells, specifying the batch of origin for each cell.
Alternatively \code{NULL} if all cells belong to the same batch.}

\item{nmads}{Numeric scalar specifying the number of median absolute deviations to be used to detect outliers.}
}
\value{
A list containing:
\itemize{
    \item \code{filter}, a logical vector of length equal to the number of cells.
    True values indicate that a cell was considered to be low quality (for any reason) and removed.
    \item \code{thresholds}, a list containing:
    \itemize{
        \item \code{sums}, a numeric vector containing the minimum threshold on the total count (in each batch, if \code{batch} is not \code{NULL}).
        \item \code{detected}, a numeric vector containing the minimum threshold on the number of detected features.
        \item \code{subsets}, a list of numeric vectors containing the maximum threshold on each feature subset proportion.
    }
    Each numeric vector is of length equal to the number of blocks (default 1 if \code{block=NULL}).
}
}
\description{
Suggest appropriate filters to be applied on the per-cell QC metrics for RNA data.
}
\examples{
# Mocking a matrix:
library(Matrix)
x <- round(abs(rsparsematrix(1000, 100, 0.1) * 100))

# Mocking up a subset:
sub <- list(Mito=rbinom(nrow(x), 1, 0.1) > 0)

# Running the analysis:
y <- initializeSparseMatrix(x)
qc <- perCellRnaQcMetrics.chan(y, sub)
filters <- suggestRnaQcFilters.chan(qc$sums, qc$detected, qc$subsets)
str(filters)

}
\author{
Aaron Lun
}
