% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignReferenceClusters.R
\name{assignReferenceClusters.chan}
\alias{assignReferenceClusters.chan}
\title{Assign cells to reference clusters}
\usage{
assignReferenceClusters.chan(
  ref,
  clusters,
  test,
  num.neighbors = 20,
  approximate = TRUE,
  num.threads = 1
)
}
\arguments{
\item{ref}{Numeric matrix for the reference dataset, where rows are dimensions and columns are cells.
This usually contains PC coordinates from \code{\link{runPCA.chan}} or friends.}

\item{clusters}{Vector or factor of length equal to \code{ncol(ref)}, containing the cluster assignment for each cell in the reference dataset.}

\item{test}{Numeric matrix for the test dataset, where rows are dimensions and columns are cells.
This usually contains PC coordinates similar to \code{ref}.}

\item{num.neighbors}{Integer scalar specifying the number of neighbors in the reference to use. 
Smaller values focus on local similarities to immediate neighbors while larger values focus on similarities to the bulk of the cluster.}

\item{approximate}{Logical scalar indicating whether an approximate neighbor search should be performed.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A data frame containing:
\itemize{
\item \code{assigned}, a vector or factor of the same type as \code{clusters}, containing cluster assignments for all cells in \code{test}.
\item \code{best.prop}, numeric vector containing the proportion of neighbors supporting the assignment of each cell in \code{assigned}.
\item \code{second.prop}, numeric vector containing the second-largest proportion of neighbors. 
}
}
\description{
Given a set of clusters in a reference dataset, assign cluster labels to cells in a test dataset using a nearest neighbor classification scheme.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
keep <- downsampleByNeighbors.chan(x, 5)
sub <- x[,keep]

labs <- iris[keep,"Species"]
full <- assignReferenceClusters.chan(sub, labs, x)
table(full$assigned, iris[,"Species"]) # compare to reference
}
\seealso{
\code{\link{downsampleByNeighbors.chan}}, which is typically used to generate a reference subsample for easier clustering before assignment of the full (test) dataset.
}
\author{
Aaron Lun
}
