% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterKmeans.R
\name{clusterKmeans.chan}
\alias{clusterKmeans.chan}
\title{K-means clustering}
\usage{
clusterKmeans.chan(
  x,
  k = 10,
  init.method = "pca-part",
  seed = 5489L,
  drop = TRUE,
  downsample = NULL,
  num.threads = 1
)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells.}

\item{k}{Integer scalar specifying the number of clusters.
This may also be a vector to perform a parameter sweep.}

\item{init.method}{String specifying the initialization method:
PCA partitioning (\code{"pca-part"}), kmeans++ (\code{"kmeans++"}) or random initialization (\code{"random"}).
This may also be a vector to perform a parameter sweep.}

\item{seed}{Integer scalar specifying the seed to use for random or kmeans++ initialization.
This may also be a vector to perform a parameter sweep.}

\item{drop}{Logical scalar indicating whether to drop the sweep-based formatting when \code{k}, \code{init.method} and \code{seed} are scalars.}

\item{downsample}{Integer scalar specifying the downsampling level (see \code{k} in \code{\link{downsampleByNeighbors.chan}}.
If \code{NULL}, no downsampling is performed.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
By default, a list is returned containing:
\itemize{
\item \code{clusters}, a factor containing the cluster assignments for each cell.
\item \code{centers}, a numeric matrix with the coordinates of the cluster centroids (dimensions in rows, centers in columns).
\item \code{iterations}, an integer scalar specifying the number of Hartigan-Wong iterations used.
\item \code{withinss}, a numeric vector containing the within-cluster sum of squares for each cluster.
}

If any of the parameters are vectors or \code{drop = FALSE}, we assume that the user is requesting a parameter sweep.
A list is returned containing \code{parameters}, a data.frame with each relevant combination of parameters;
and \code{results}, a list of length equal to the number of rows of \code{parameters}, where each entry contains the result for the corresponding parameter combination.

If \code{downsample} is not \code{NULL}, only the \code{clusters} element of the list will be reported.
}
\description{
Perform k-means clustering using kmeans++ initialization with the Hartigan-Wong algorithm.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
clustering <- clusterKmeans.chan(x)
table(clustering$clusters)

# Parameter sweep mode.
swept <- clusterKmeans.chan(x, k=c(2,5,10), 
    init.method=c("pca-part", "random"), 
    seed=c(1,2,3))
swept$parameters
length(swept$results)
table(swept$results[[1]]$clusters)

}
\author{
Aaron Lun
}
