% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnnCorrect.R
\name{mnnCorrect.chan}
\alias{mnnCorrect.chan}
\title{MNN correction}
\usage{
mnnCorrect.chan(x, batch, k = 15, nmads = 3)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells,
typically generated from \code{\link{runPCA.chan}}.}

\item{batch}{Vector or factor of length equal to the number of cells, specifying the batch of origin for each cell.
Alternatively \code{NULL} if all cells belong to the same batch.}

\item{k}{Integer scalar specifying the number of neighbors to use when identifying MNN pairs.}

\item{nmads}{Numeric scalar specifying the number of MADs to use when computing correction vectors.}
}
\value{
List containing:
\itemize{
\item \code{corrected}, a numeric matrix of the same dimensions as \code{x}, containing the corrected values.
\item \code{merge.order}, character vector containing the unique levels of \code{batch} in the automatically determined merge order.
The first batch is used as the reference and all other batches are iteratively merged.
\item \code{num.pairs}, integer vector of length equal to the number of batches minus 1.
This contains the number of MNN pairs at each merge.
}
}
\description{
Apply mutual nearest neighbor (MNN) correction to remove batch effects from a low-dimensional matrix.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
b <- sample(1:2, ncol(x), replace=TRUE)
corrected <- mnnCorrect.chan(x, b)
str(corrected)

}
\author{
Aaron Lun
}
