% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelGeneVar.R
\name{modelGeneVar.chan}
\alias{modelGeneVar.chan}
\title{Model per-gene variance}
\usage{
modelGeneVar.chan(x, span = 0.3, num.threads = 1)
}
\arguments{
\item{x}{A list containing the log-expression matrix, typically produced by \code{\link{logNormCounts.chan}}.}

\item{span}{Numeric scalar containing the span to use for LOWESS smoothing of the trend.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A data.frame of statistics for each gene in \code{x}.
}
\description{
Model the per-gene variance in log-expression values, accounting for the mean-variance trend.
}
\examples{
library(Matrix)
x <- round(abs(rsparsematrix(1000, 100, 0.1) * 100))
y <- initializeSparseMatrix(x)
normed <- logNormCounts.chan(y)

df <- scran.chan:::modelGeneVar.chan(normed)
df[1:10,]
plot(df$means, df$variances)
points(df$means, df$fitted, col="red", pch=16, cex=0.5)

}
\author{
Aaron Lun
}
