% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelGeneVar.R
\name{modelGeneVar.chan}
\alias{modelGeneVar.chan}
\title{Model per-gene variance}
\usage{
modelGeneVar.chan(
  x,
  span = 0.3,
  use.fixed.width = FALSE,
  fixed.width = 1,
  fixed.min.count = 200,
  num.threads = 1,
  batch = NULL
)
}
\arguments{
\item{x}{A list containing the log-expression matrix, typically produced by \code{\link{logNormCounts.chan}}.}

\item{span}{Numeric scalar containing the span to use for LOWESS smoothing of the trend.}

\item{use.fixed.width}{Logical scalar specifying whether to apply a fixed-width constraint on the smoothing,
to avoid problems with large differences in density across the range of means.}

\item{fixed.width}{Numeric scalar specifying the window width when `use.fixed.width = TRUE`.}

\item{fixed.min.count}{Integer scalar specifying the minimum number of genes in a window when `use.fixed.width = TRUE`.
Windows with too few genes are expanded until they contain the required minimum.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}

\item{batch}{Vector or factor of length equal to the number of cells, specifying the batch of origin for each cell.
Alternatively \code{NULL} if all cells belong to the same batch.}
}
\value{
A list containing \code{statistics}, a data frame of statistics for each gene in \code{x}.

If \code{batch} is supplied, this list will also contain \code{per.batch}, a list of data frames with statistics for each batch.
(In that case, \code{statistics} contains the averaged statistics for each gene across batches.)
}
\description{
Model the per-gene variance in log-expression values, accounting for the mean-variance trend.
}
\examples{
library(Matrix)
x <- round(abs(rsparsematrix(1000, 100, 0.1) * 100))
y <- initializeSparseMatrix(x)
normed <- logNormCounts.chan(y)

dfl <- modelGeneVar.chan(normed)
df <- dfl$statistics
df[1:10,]
plot(df$means, df$variances)
points(df$means, df$fitted, col="red", pch=16, cex=0.5)

}
\author{
Aaron Lun
}
