% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSNE.R
\name{runTSNE.chan}
\alias{runTSNE.chan}
\title{Run an approximate t-SNE}
\usage{
runTSNE.chan(x, perplexity = 30, num.threads = 1)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells.}

\item{perplexity}{Numeric scalar specifying the perplexity to use in the t-SNE algorithm.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A numeric matrix where rows are cells and columns are the two dimensions of the embedding.
}
\description{
Compute the t-stochastic neighbor embedding with assorted approximations for speed.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
embedding <- runTSNE.chan(x)
plot(embedding[,1], embedding[,2], col=iris[,5])

}
\author{
Aaron Lun
}
