% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSNE.R
\name{runTSNE.chan}
\alias{runTSNE.chan}
\title{Run an approximate t-SNE}
\usage{
runTSNE.chan(
  x,
  perplexity = 30,
  interpolate = -1,
  max.depth = 7,
  seed = 42,
  drop = TRUE,
  num.threads = 1
)
}
\arguments{
\item{x}{Numeric matrix where rows are dimensions and columns are cells.}

\item{perplexity}{Numeric scalar specifying the perplexity to use in the t-SNE algorithm.
This may also be a vector to perform a parameter sweep.}

\item{interpolate}{Integer scalar specifying the grid resolution for interpolating repulsive forces (larger is slower but more accurate).
A value of zero disables interpolation, while a value of -1 only uses interpolation for large datasets.
This may also be a vector to perform a parameter sweep.}

\item{max.depth}{Integer scalar specifying the maximum depth of the Barnes-Hut quad trees (larger is slower but more accurate).
Set to a large integer (e.g., 1000) to eliminate depth restrictions.
This may also be a vector to perform a parameter sweep.}

\item{seed}{Integer scalar specifying the seed to use for generating the initial coordinates.
This may also be a vector to perform a parameter sweep.}

\item{drop}{Logical scalar indicating whether to drop the sweep-based formatting when \code{perplexity}, \code{interpolate} and \code{max.depth} are scalars.}

\item{num.threads}{Integer scalar specifying the number of threads to use.}
}
\value{
By default, a numeric matrix where rows are cells and columns are the two dimensions of the embedding.

If any of the parameters are vectors or \code{drop = FALSE}, we assume that the user is requesting a parameter sweep.
A list is returned containing \code{parameters}, a data.frame with each relevant combination of parameters;
and \code{results}, a list of numeric matrices of length equal to the number of rows of \code{parameters}, containing the embeddings for each combination.
}
\description{
Compute the t-stochastic neighbor embedding with assorted approximations for speed.
}
\examples{
x <- t(as.matrix(iris[,1:4]))
embedding <- runTSNE.chan(x)
plot(embedding[,1], embedding[,2], col=iris[,5])

# Parameter sweep mode.
swept <- runTSNE.chan(x, 
    perplexity=c(10, 20, 50),
    seed=c(42, 88))
swept$parameters
head(swept$results[[1]])

}
\author{
Aaron Lun
}
